# docs/03-api-contract-openapi.yaml
openapi: 3.0.3
info:
  title: BNC Financial Flow API
  version: 1.0.0
  description: |
    API para fluxo de caixa (realizado e projetado), contas a receber/pagar,
    multi-moeda e operações idempotentes de pagamento.
servers:
  - url: https://api.seudominio.com
security:
  - bearerAuth: []
...
paths:
  /v1/accounts:
    post:
      tags: [Accounts]
      summary: Criar conta financeira
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Conta criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
    get:
      tags: [Accounts]
      summary: Listar contas
      operationId: listAccounts
      responses:
        '200':
          description: Lista de contas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccountResponse'
  /v1/receivables:
    post:
      tags: [Receivables]
      summary: Criar contas a receber com parcelamento automático
      operationId: createReceivable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReceivableRequest'
      responses:
        '201':
          description: Recebível criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceivableResponse'
    get:
      tags: [Receivables]
      summary: Listar contas a receber
      operationId: listReceivables
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReceivableStatus'
      responses:
        '200':
          description: Lista de recebíveis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReceivablesResponse'
...
  /v1/installments/{installmentId}/pay:
    post:
      tags: [Installments]
      summary: Pagar parcela (idempotente)
      operationId: payInstallment
      parameters:
        - $ref: '#/components/parameters/InstallmentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayInstallmentRequest'
      responses:
        '200':
          description: Parcela paga
          headers:
            Idempotency-Replayed:
              schema:
                type: boolean
              description: true quando a resposta veio do cache idempotente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResultResponse'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
...
  /v1/payables/{payableId}/pay:
    post:
      tags: [Payables]
      summary: Pagar conta a pagar (idempotente)
      operationId: payPayable
      parameters:
        - $ref: '#/components/parameters/PayableId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayPayableRequest'
      responses:
        '200':
          description: Conta a pagar quitada
          headers:
            Idempotency-Replayed:
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResultResponse'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 120
...
    InstallmentId:
      name: installmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PayableId:
      name: payableId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    IdempotencyConflict:
      description: Mesma chave idempotente usada com payload diferente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessRuleViolation:
      description: Regra de negócio violada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    AccountType:
      type: string
      enum: [BANK, CASH, WALLET, OTHER]
    ReceivableStatus:
      type: string
      enum: [OPEN, PARTIALLY_PAID, PAID, OVERDUE, CANCELED]
...
    CreateAccountRequest:
      type: object
      required: [name, type, baseCurrency]
      properties:
        name:
          type: string
          maxLength: 100
        type:
          $ref: '#/components/schemas/AccountType'
        baseCurrency:
          type: string
          pattern: '^[A-Z]{3}$'
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        baseCurrency:
          type: string
    CreateReceivableRequest:
      type: object
      required:
        [customerName, totalAmount, currency, saleDate, installmentsCount]
      properties:
        customerName:
          type: string
          maxLength: 180
        description:
          type: string
        totalAmount:
          type: number
          format: double
          minimum: 0.01
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        saleDate:
          type: string
          format: date
        installmentsCount:
          type: integer
          minimum: 1
          maximum: 120
    ReceivableResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ReceivableStatus'
        customerName:
          type: string
        totalAmount:
          type: number
        currency:
          type: string
        installmentsCount:
          type: integer
...
    PayInstallmentRequest:
      type: object
      required: [accountId, paidAt]
      properties:
        accountId:
          type: string
          format: uuid
        paidAt:
          type: string
          format: date-time
        exchangeRateId:
          type: string
          format: uuid
          nullable: true
        exchangeRate:
          type: number
          nullable: true
          minimum: 0.00000001
        description:
          type: string
    PayPayableRequest:
      type: object
      required: [accountId, paidAt]
      properties:
        accountId:
          type: string
          format: uuid
        paidAt:
          type: string
          format: date-time
        exchangeRateId:
          type: string
          format: uuid
          nullable: true
        exchangeRate:
          type: number
          nullable: true
          minimum: 0.00000001
        description:
          type: string
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        direction:
          type: string
          enum: [IN, OUT]
        amountOriginal:
          type: number
        currencyOriginal:
          type: string
        exchangeRate:
          type: number
        amountConverted:
          type: number
        currencyConverted:
          type: string
        occurredAt:
          type: string
          format: date-time
        originType:
          type: string
        originId:
          type: string
          format: uuid
          nullable: true
        reversalOfTransactionId:
          type: string
          format: uuid
          nullable: true
    PaymentResultResponse:
      type: object
      properties:
        installmentId:
          type: string
          format: uuid
          nullable: true
        payableId:
          type: string
          format: uuid
          nullable: true
        transaction:
          $ref: '#/components/schemas/TransactionResponse'
...
    PaginatedReceivablesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReceivableResponse'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: IDEMPOTENCY_CONFLICT
        message:
          type: string
        details:
          type: object
          additionalProperties: true